fix: spell checking fix and extract buttons theme
